adding the delete method and test it succsesfully with postman

+++ controllers/blogs.js
@@ -37,12 +37,12 @@ blogsRouter.post('/', async (request, response) => {
     return response.status(401).json({ error: 'token invalid' })
   }
   const user = await User.findById(decodedToken.id)
-  // if (!body.likes) {
-  //   return response.status(400).json({ error: 'likes are required' })
-  // }
-  // if (!body.title || !body.url) {
-  //   return response.status(400).json({ error: 'Title or URL is missing' })
-  // }
+  if (!body.likes) {
+    return response.status(400).json({ error: 'likes are required' })
+  }
+  if (!body.title || !body.url) {
+    return response.status(400).json({ error: 'Title or URL is missing' })
+  }
 
   const blog = new Blog({
     title: body.title,
@@ -67,6 +67,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'only the user who added the blog can delete it' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
